Remove private crate
