+++ Cargo.toml
@@ -9,8 +9,13 @@ readme = "README.md"
 keywords = ["array", "macro", "initialization", "init"]
 license = "MIT OR Apache-2.0"
 
-[dependencies]
-arr_macro_impl = { version = "0.1.3", path = "impl" }
+[lib]
+proc-macro = true
 
-[workspace]
-members = ["impl"]
+[dependencies]
+quote = "1.0"
+
+[dependencies.syn]
+version = "1.0"
+default-features = true
+features = ["full"]
