--- impl/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "arr_macro_impl"
-version = "0.1.3"
-authors = ["Josh Mcguigan"]
-edition = "2021"
-description = "Private impl crate for arr_macro"
-repository = "https://github.com/JoshMcguigan/arr_macro"
-license = "MIT OR Apache-2.0"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1.0"
-
-[dependencies.syn]
-version = "1.0"
-default-features = true
-features = ["full"]
